+++ main.go
@@ -24,6 +24,7 @@ type config struct {
 
 	PollingIntervalSeconds int `yaml:"polling_interval_seconds"`
 	Projects               []project
+	Wildcards              []wildcard
 }
 
 type client struct {
@@ -33,7 +34,16 @@ type client struct {
 
 type project struct {
 	Name string
-	Ref  string
+	Refs []string
+}
+
+type wildcard struct {
+	Search string
+	Owner  struct {
+		Name string
+		Kind string
+	}
+	Refs []string
 }
 
 var (
@@ -83,36 +93,122 @@ func (c *client) getProject(name string) *gitlab.Project {
 	return p
 }
 
+func (c *client) fetchProjectsFromWildcards() {
+	for _, w := range c.config.Wildcards {
+		for _, p := range c.listProjects(&w) {
+			c.config.Projects = append(c.config.Projects, p)
+		}
+	}
+}
+
+func (c *client) listProjects(w *wildcard) (projects []project) {
+	log.Printf("-> Listing all projects using search pattern : '%s' with owner '%s' (%s)", w.Search, w.Owner.Name, w.Owner.Kind)
+
+	trueVal := true
+	falseVal := false
+	var gps []*gitlab.Project
+	var err error
+
+	switch w.Owner.Kind {
+	case "user":
+		gps, _, err = c.Projects.ListUserProjects(
+			w.Owner.Name,
+			&gitlab.ListProjectsOptions{
+				Archived: &falseVal,
+				Simple:   &trueVal,
+				Search:   &w.Search,
+			},
+		)
+	case "group":
+		gps, _, err = c.Groups.ListGroupProjects(
+			w.Owner.Name,
+			&gitlab.ListGroupProjectsOptions{
+				Archived: &falseVal,
+				Simple:   &trueVal,
+				Search:   &w.Search,
+			},
+		)
+	default:
+		log.Fatalf("Invalid owner kind '%s' must be either 'user' or 'group'", w.Owner.Kind)
+		os.Exit(1)
+	}
+
+	if err != nil {
+		log.Fatalf("Unable to list projects with search pattern '%s' from the GitLab API : %v", w.Search, err.Error())
+		os.Exit(1)
+	}
+
+	for _, gp := range gps {
+		log.Printf("-> Found project : %s", gp.PathWithNamespace)
+		projects = append(
+			projects,
+			project{
+				Name: gp.PathWithNamespace,
+				Refs: w.Refs,
+			},
+		)
+	}
+
+	return
+}
+
 func (c *client) pollProject(p project) {
-	gp := c.getProject(p.Name)
-	log.Printf("--> Polling ID: %v | %v:%v", gp.ID, p.Name, p.Ref)
+	for _, ref := range p.Refs {
+		go c.pollProjectRef(p.Name, ref)
+	}
+}
+
+func (c *client) pollProjectRef(name, ref string) {
+	gp := c.getProject(name)
+	log.Printf("--> Polling ID: %v | %v:%v", gp.ID, name, ref)
 
 	var lastPipeline *gitlab.Pipeline
 
 	for {
-		pipelines, _, _ := c.Pipelines.ListProjectPipelines(gp.ID, &gitlab.ListProjectPipelinesOptions{Ref: gitlab.String(p.Ref)})
+		pipelines, _, _ := c.Pipelines.ListProjectPipelines(gp.ID, &gitlab.ListProjectPipelinesOptions{Ref: gitlab.String(ref)})
 		if lastPipeline == nil || lastPipeline.ID != pipelines[0].ID || lastPipeline.Status != pipelines[0].Status {
 			if lastPipeline != nil {
-				runCount.WithLabelValues(p.Name, p.Ref).Inc()
+				runCount.WithLabelValues(name, ref).Inc()
 			}
 
-			lastPipeline, _, _ = c.Pipelines.GetPipeline(gp.ID, pipelines[0].ID)
+			if len(pipelines) > 0 {
+				lastPipeline, _, _ = c.Pipelines.GetPipeline(gp.ID, pipelines[0].ID)
+				lastRunDuration.WithLabelValues(name, ref).Set(float64(lastPipeline.Duration))
 
-			lastRunDuration.WithLabelValues(p.Name, p.Ref).Set(float64(lastPipeline.Duration))
-
-			for _, s := range []string{"success", "failed", "running"} {
-				if s == lastPipeline.Status {
-					status.WithLabelValues(p.Name, p.Ref, s).Set(1)
-				} else {
-					status.WithLabelValues(p.Name, p.Ref, s).Set(0)
+				for _, s := range []string{"success", "failed", "running"} {
+					if s == lastPipeline.Status {
+						status.WithLabelValues(name, ref, s).Set(1)
+					} else {
+						status.WithLabelValues(name, ref, s).Set(0)
+					}
 				}
+			} else {
+				log.Printf("Could not find any pipeline for %s:%s", name, ref)
 			}
 		}
 
-		timeSinceLastRun.WithLabelValues(p.Name, p.Ref).Set(float64(time.Since(*lastPipeline.CreatedAt).Round(time.Second).Seconds()))
+		if lastPipeline != nil {
+			timeSinceLastRun.WithLabelValues(name, ref).Set(float64(time.Since(*lastPipeline.CreatedAt).Round(time.Second).Seconds()))
+		}
 
 		time.Sleep(time.Duration(c.config.PollingIntervalSeconds) * time.Second)
 	}
+}
+
+func (c *client) pollProjects() {
+	c.fetchProjectsFromWildcards()
+	log.Printf("-> %d project(s) configured with a total of %d ref(s)", len(c.config.Projects), c.sumTotalRefs())
+	for _, p := range c.config.Projects {
+		c.pollProject(p)
+	}
+}
+
+func (c *client) sumTotalRefs() (refs int) {
+	refs = 0
+	for _, p := range c.config.Projects {
+		refs += len(p.Refs)
+	}
+	return
 }
 
 func init() {
@@ -139,14 +235,13 @@ func main() {
 		os.Exit(1)
 	}
 
-	if len(config.Projects) < 1 {
-		log.Fatalf("You need to configure at least one project/ref to poll, none given")
+	if len(config.Projects) < 1 && len(config.Wildcards) < 1 {
+		log.Fatalf("You need to configure at least one project/wildcard to poll, none given")
 		os.Exit(1)
 	}
 
 	log.Printf("-> Starting exporter")
 	log.Printf("-> Polling %v every %vs", config.Gitlab.URL, config.PollingIntervalSeconds)
-	log.Printf("-> %v project(s) configured", len(config.Projects))
 
 	c := &client{
 		gitlab.NewClient(nil, config.Gitlab.Token),
@@ -154,10 +249,7 @@ func main() {
 	}
 
 	c.SetBaseURL(config.Gitlab.URL)
-
-	for _, p := range config.Projects {
-		go c.pollProject(p)
-	}
+	c.pollProjects()
 
 	// Configure liveness and readiness probes
 	health := healthcheck.NewHandler()